+++ src/sections/index.ts
@@ -1,2 +1,6 @@
 export * from './Wellcome';
 export * from './Challenge';
+export * from './Contact';
+export * from './Projects';
+export * from './AboutMe';
+export * from './Skills';
